feat: add sidebar links and user panel
