+++ app/(2. pages)/panel/layout.tsx
@@ -8,7 +8,7 @@ const Layout = (): ReactNode => {
     <AppShell
       padding="md"
       navbar={
-        <Navbar width={{ base: 300 }} height={500} p="xs">
+        <Navbar width={{ base: 300 }} className="min-h-fit" p="xs">
           <PanelNavbar />
         </Navbar>
       }
