+++ app/5. entities/sidebar-links/model.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export interface MainLinkProps {
+  icon: React.ReactNode;
+  color: string;
+  label: string;
+}
