--- app/5. entities/sidebar/main-links.tsx
@@ -1,45 +1,0 @@
-import React, { ReactNode } from "react";
-import { IconGitPullRequest, IconAlertCircle, IconMessages, IconDatabase } from "@tabler/icons-react";
-import { ThemeIcon, UnstyledButton, Group, Text, MantineTheme, CSSObject } from "@mantine/core";
-
-interface MainLinkProps {
-  icon: React.ReactNode;
-  color: string;
-  label: string;
-}
-
-function MainLink({ icon, color, label }: MainLinkProps): ReactNode {
-  return (
-    <UnstyledButton
-      sx={(theme: MantineTheme): CSSObject => ({
-        display: "block",
-        width: "100%",
-        padding: theme.spacing.xs,
-        borderRadius: theme.radius.sm,
-        color: theme.colorScheme === "dark" ? theme.colors.dark[0] : theme.black,
-        "&:hover": {
-          backgroundColor: theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.gray[0]
-        }
-      })}
-    >
-      <Group>
-        <ThemeIcon color={color} variant="light">
-          {icon}
-        </ThemeIcon>
-        <Text size="sm">{label}</Text>
-      </Group>
-    </UnstyledButton>
-  );
-}
-
-const data = [
-  { icon: <IconGitPullRequest size="1rem" />, color: "blue", label: "Pull Requests" },
-  { icon: <IconAlertCircle size="1rem" />, color: "teal", label: "Open Issues" },
-  { icon: <IconMessages size="1rem" />, color: "violet", label: "Discussions" },
-  { icon: <IconDatabase size="1rem" />, color: "grape", label: "Databases" }
-];
-
-export function MainLinks(): ReactNode {
-  const links = data.map(link => <MainLink {...link} key={link.label} />);
-  return <div>{links}</div>;
-}
