+++ app/5. entities/user-sidebar/index.tsx
@@ -26,10 +26,8 @@ export function UserSidebar(): ReactNode {
         }}
       >
         <Group>
-          <Avatar
-            src="https://images.unsplash.com/photo-1508214751196-bcfd4ca60f91?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=255&q=80"
-            radius="xl"
-          />
+          {/* todo: if user have avatar fill src attr */}
+          <Avatar src="" radius="xl" />
           <Box sx={{ flex: 1 }}>
             <Text size="sm" weight={500}>
               Amy Horsefighter
@@ -38,7 +36,6 @@ export function UserSidebar(): ReactNode {
               ahorsefighter@gmail.com
             </Text>
           </Box>
-
           {theme.dir === "ltr" ? <IconChevronRight size={rem(18)} /> : <IconChevronLeft size={rem(18)} />}
         </Group>
       </UnstyledButton>
